chore: move commands to their own directories
