+++ src/cli/commands/index.ts
@@ -1,4 +1,4 @@
-export * from './build';
-export * from './compile';
-export * from './info';
-export * from './migrate';
+export * from './build/build';
+export * from './compile/compile';
+export * from './info/info';
+export * from './migrate/migrate';
