+++ src/cli/commands/info/info.ts
